refactor(csr): merge binary_indexed_mm into binary_indexed; split indexed CUDA ops
- Merge binary_csrmm_indexed (function, kernels, autodiff, primitive) and its
  tests into binary_indexed.py / binary_indexed_test.py; drop binary_indexed_mm.py
  and binary_indexed_mm_test.py.
- Extract the indexed (perm-hetero) CUDA kernels out of binary_csrmv.cu and
  binary_csrmm.cu into dedicated binary_indexed_csrmv.cu and
  binary_indexed_csrmm.cu. The homogeneous indexed path still reuses the plain
  _homo kernels in the original files.
- Update _csr/__init__.py and _csr/main.py imports accordingly.
